Fix crash
The crash would occur when updating the todo cells after an item had been removed from the today list in Things.app.

There seems to still be a remnant of the todo object, but there is no title, so that is what we check for. (I this is because of the async call to update the agenda.)